Move common functions to transactron.
