+++ test/common/functions.py
@@ -1,14 +1,10 @@
 from amaranth import *
 from typing import TYPE_CHECKING
-from coreblocks.utils import LayoutLike
 from transactron._utils import RecordValueDict, RecordIntDict
+
 
 if TYPE_CHECKING:
     from .wrappers import TestGen
-
-
-def data_layout(val: int) -> LayoutLike:
-    return [("data", val)]
 
 
 def set_inputs(values: RecordValueDict, field: Record) -> "TestGen[None]":
@@ -30,60 +26,3 @@ def get_outputs(field: Record) -> "TestGen[RecordIntDict]":
         else:  # field is a Record
             result[name] = yield from get_outputs(val)
     return result
-
-
-def neg(x: int, xlen: int) -> int:
-    """
-    Computes the negation of a number in the U2 system.
-
-    Parameters
-    ----------
-    x: int
-        Number in U2 system.
-    xlen : int
-        Bit width of x.
-
-    Returns
-    -------
-    return : int
-        Negation of x in the U2 system.
-    """
-    return (-x) & (2**xlen - 1)
-
-
-def int_to_signed(x: int, xlen: int) -> int:
-    """
-    Converts a Python integer into its U2 representation.
-
-    Parameters
-    ----------
-    x: int
-        Signed Python integer.
-    xlen : int
-        Bit width of x.
-
-    Returns
-    -------
-    return : int
-        Representation of x in the U2 system.
-    """
-    return x & (2**xlen - 1)
-
-
-def signed_to_int(x: int, xlen: int) -> int:
-    """
-    Changes U2 representation into Python integer
-
-    Parameters
-    ----------
-    x: int
-        Number in U2 system.
-    xlen : int
-        Bit width of x.
-
-    Returns
-    -------
-    return : int
-        Representation of x as signed Python integer.
-    """
-    return x | -(x & (2 ** (xlen - 1)))
